Move components to embeds
