+++ lenny/embeds/config/sources.py
@@ -1,7 +1,6 @@
 import discord
 
-from components.items import BaseSeparator
-from components.paginated_view import PaginatedLayoutView
+from embeds.components import BaseSeparator, PaginatedLayoutView
 from embeds.config.config import ConfigAllowButton
 from logic.config import Config
 from logic.dnd.source import ContentChoice, Source, SourceList
